requirements.txt
django==1.8.1
dj-database-url
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
requests==2.9.1
whitenoise
djangorestframework
